pyproject.toml
[tool.poetry]
name = "investiq"
version = "0.1.0"
description = ""
authors = ["Delphine Tai-Beauchamp <delphinetaibeauchamp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
tweepy = "^4.14.0"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
pytest = "^8.3.3"
openai = "^1.52.0"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
jinja2 = "^3.1.4"
pydantic = "^2.9.2"
langchain = "^0.3.4"
langchain-openai = "^0.2.3"
streamlit = "^1.39.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
